refactor app.js into components
